Update README.md + Change Access Modifiers

+++ src/main/java/uet/oop/space_invaders/spaceinvaders/BossEnemy.java
@@ -15,8 +15,8 @@ public class BossEnemy extends Enemy {
     private int health;
 
     // Hitbox dimensions
-    private static final int WIDTH = 146;
-    private static final int HEIGHT = 105;
+    private static final double WIDTH = 146;
+    private static final double HEIGHT = 105;
 
     // Horizontal movement speed
     private double horizontalSpeed;
